spv: Move initial rescan to global syncer loop
This moves the initial account and address discovery and rescan from the
per-peer startup procedure to the global syncer startup sync procedure.

This makes the code easier to reason about and removes the need for the
atomic lock to the discoverAccounts field.

The previous code would execute a rescan after every connected peer, if
it had any new headers that connected to the main chain.  With the
changes in this commit, the syncer now only performs the full rescan
once at startup and relies on the header announcement logic to scan when
new blocks are relayed from peers after the initial sync.