multi: Refactor wallet.NetworkBackend.
This unifies the wallet's Peer and NetworkBackend interfaces into a
single interface, given the SPV peer syncer now fulfills both roles.

In order to complete the refactor, the following changes are made:

- The funcs that were in Peer are moved to NetworkBackend.
- The missing funcs are added to chain.Syncer, forwading the calls to
  the dcrd client connection.
- chain.Syncer assigns itself as NetworkBackend to the wallet, instead
  of assigning the client connection.
- deployments package is modified to take a more limited interface
  (Querier) instead of the syncer var directly, to avoid circular pkg
  dependencies.
- The live ticket querier (in tickets.go) is modified to take a more
  limited interface instead of the syncer var directly, to avoid
  circular pkg dependencies.
- Call sites that type selected on *dcrd.RPC are modified to type select
  on *chain.Syncer.
- A Synced() function is added to the NetworkBackend interface so that
  the JSON-RPC server can more accurately query for synced status from
  the underlying syncer.

In the future, the dependency on concrete type checking could be further
reduced on other call sites by using the same technique that was
employed in this commit for the deployments package.