chain: Wait for dcrd to sync before starting RPC sync
This prevents the wallet from syncing from blockConnected messages when
it is connected to a dcrd instance that is not yet as synced as the one
the wallet has used in the past. This improves performance by batching
the sync work in the initial getHeaders stage and prevents wallet misuse
by ensuring the underlying dcrd is also synced to recent blocks.

Note that the the case of a new wallet (with recorded tip height 0)
connected to a network isolated dcrd instance (header and block height
both 0) is explicitly allowed to proceed without waiting. This handles
the use-cases of airgapped and empty simnet wallets.