Implement spendoutputs JSON-RPC method
This method provides greater coin control when creating and sending
transactions.  Rather than the wallet performing UTXO selection on its own,
only the specified inputs and outputs (with the exception of wallet change)
will be used.

This is intended to be a more ergonomic and easier to use alternative to the
raw transaction RPCs (createrawtransaction, signrawtransaction, and
sendrawtransaction).