Introduce VSPTicket struct.
Upon creation, the new VSPTicket struct is initialized with all of the
values necessary to register a ticket with a VSP. It is then passed into
the VSP client. This helps to ensure no invalid tickets are added to a
VSP client, thus removing a bunch of error handling from vsp.go and
feepayment.go. It also reduces database access because values such as
the raw transactions are retrieved once and then held in the VSPTicket
struct to be reused.

As a result, a lot of duplicated logic and complexity are removed from
vsp.go and feepayment.go, and a lot of logic specific to VSP tickets is
removed from wallet.go and into vspticket.go.