multi: Expose disablerelaytx
This setting allows the wallet to not sync mempool transactions when in
SPV mode. This is useful to reduce cpu, memory and bandwidth consumption
for wallets that see very few useful mempool transactions, such as
wallets meant for only occasional use or mobile wallets.

The setting is exposed as a CLI or config argument --spvdisablerelaytx.