chain: Wait for errgroup before returning from sync
If syncing over JSON-RPC is being reattempted in a loop, as the main package
does, not waiting for the errgroup that runs the wallet's mixing client may
cause it to concurrently enter mixclient.Client.Run again before it has
returned.  This eventually results in a panic due to closing the already-closed
warming chan.

This is a backport candidate for 2.0.1.