Remove Stakepool mode.
Stakepool mode was enabled using the `--stakepoolcoldextkey` flag which
has been removed, and with dcrwallet unable to operate in this mode the
`stakepooluserinfo` and `ticketsforaddress` RPCs become redundant so
have also been removed.

RPC server version bumped to 10.0.0.

All functionality which previously depended on the StakeStore has been
removed, therefore the StakeStore itself can be removed.

Newly created databases will no longer be initialized with the buckets
used by StakeStore, and a database upgrade which changed the data format
in one of those buckets has been updated so it no longer attempts to
make that change.