Move VSP client code to wallet package.
This moves the VSP client code from its own package into the wallet package.
Moving this code should have minimal impact on consumers of the vsp package
because they will have been importing the wallet package already due to the fact
that a VSP client could not be created without an instance of wallet.

The benefits of this are:

- vsp client creation now exists only in the wallet package, rather than being
  scattered across 3 packages.
- vsp client loading funcs and variables are no longer package global.
- simplifies the creation of a vsp client because the wallet no longer needs to
  be passed into the func as a param.
- no longer necessary to abstract the vsp client away by referring to its
  methods, the client can just be used directly.